refactor(vm): lift the durable-write primitive out of the sandbox cluster
`template_catalog` reached backward into `sandbox::persistence` for
`atomic_write`, which is why `sandbox.rs` carried a `pub(crate) use` with
a clippy allow explaining that a `pub(super)` item cannot be re-exported
wider. The dependency is the wrong way round: writing a file durably is
not sandbox-lifecycle knowledge, and the catalog is the module the
charter moves out of this crate first.

It becomes `crate::atomic_file`, with the two entry points named for what
they do rather than how they nest: `write` for callers that treat an
unconfirmed directory `fsync` as an error, and `write_reporting_durability`
for the sandbox record store, which reports it as a warning and keeps the
record. That distinction was the only reason the inner function existed;
now it says so.

No behavior change — same temp-sibling-fsync-rename-fsync sequence, same
error mapping. 219 unit tests unchanged.
