+++ virt/arcbox-vm/src/template_catalog.rs
@@ -23,8 +23,8 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use tracing::warn;
 
+use crate::atomic_file;
 use crate::error::{Result, VmmError};
-use crate::sandbox::atomic_write;
 
 /// Directory under the vmm data dir holding one JSON file per template name.
 const CATALOG_DIR: &str = "template-catalog";
@@ -479,7 +479,7 @@ impl TemplateCatalog {
         })
         .map_err(VmmError::Io)?;
         let bytes = serde_json::to_vec_pretty(record)?;
-        atomic_write(&self.record_path(&record.name), &bytes)
+        atomic_file::write(&self.record_path(&record.name), &bytes)
     }
 
     fn remove_record(&self, name: &str) -> Result<()> {
